backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.0
httpx>=0.27.0
pydantic>=2.8.0
